Changes: * Rename module 'data_sources.py' to 'api.py' to better clarify its purpose * Add module 'utils.py' which contains functions used by the main app, to avoid cluttering the main app with auxiliar functions.
